requirements.txt
streamlit
openai
requests
yfinance
matplotlib
pandas
Pillow
PyMuPDF
